ANDAUTH-46 Synchronize more informations from the profile * Refactoring
